+++ include/BriandDefines.hxx
@@ -179,7 +179,7 @@
         #define STACK_MgrInst 5120
         #define STACK_TorProxy 2816
         #define STACK_TorProxyReq 4096
-        #define STACK_StreamWR 3072
+        #define STACK_StreamWR 3584
         #define STACK_StreamRD 3072
     #else
         #define STACK_TorEsp32 4096