+++ NEWS.md
@@ -23,9 +23,10 @@
 * Removed parameter *Component config -> FreeRTOS -> ENABLED Place FreeRTOS functions into Flash* I think causing cache failures and frequent download.
 * Better cache access with waiting search requests
 * Better stack sizes when log suppressed
-
-* Possible bug in stream: after some cells exchanged, seems unrecognize all. Maybe due to the additional bytes download?
 * Possible bug in tor relay searcher: too much rebuild cache requests... SOLVED!
+* Possible bug in stream: after some cells exchanged, seems unrecognize all. Maybe due to the additional bytes download?
+* Separate circuit log from stream log to debug better
+* Socks5 Proxy better task management (simplified, using also make_shared to avoid invalid reads)
 
 ## 2021-08-18